Refactor registry file names and update badge status
